+++ CHANGELOG.md
@@ -55,6 +55,21 @@
     * `ECCurve.coordinateLengthBytes`
     * `ECCurve.signatureLengthBytes`
 
+* HMAC Support
+* Symmetric Encryption
+    * Supported Algorithms
+        * AES
+            * GCM
+            * CBC-HMAC
+            * CBC
+            * ECB
+            * KW
+        * ChaCha-Poly1305
+    * Add algorithm mappings to indispensable-josef **This is a binary-incompatible change**
+        * `ivLength` and `encryptionKeyLength` now return `BitLength` instead of `Int`
+        * `text` is now properly called `identifier`
+* Moved `HazardousMaterials` annotation from `supreme` to `indispensable`
+
 ### 3.14.0 (Supreme 0.7.0)
 
 * Certificate Improvements: